ldap_sync/requirements.txt
PyYAML==6.0
requests==2.28.2
python-ldap==3.4.3
urllib3==1.26.14
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
